+++ command/gii/simpleWorkflow/SimpleWorkflowGenerator.php
@@ -1,0 +1,6 @@
+<?php
+class SimpleWorkflowGenerator extends CCodeGenerator
+{
+    public $codeModel='application.gii.simpleWorkflow.SimpleWorkflowCode';
+}
+?>